chore: Move processing annotation to core module
