+++ vss-processor/src/main/kotlin/org/eclipse/kuksa/vssprocessor/spec/VssSpecificationSpecModel.kt
@@ -13,12 +13,12 @@ import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.asTypeName
-import org.eclipse.kuksa.vsscore.model.VssNode
-import org.eclipse.kuksa.vsscore.model.VssProperty
-import org.eclipse.kuksa.vsscore.model.VssSpecification
-import org.eclipse.kuksa.vsscore.model.className
-import org.eclipse.kuksa.vsscore.model.name
-import org.eclipse.kuksa.vsscore.model.variableName
+import org.eclipse.kuksa.vsscore.model.model.VssNode
+import org.eclipse.kuksa.vsscore.model.model.VssProperty
+import org.eclipse.kuksa.vsscore.model.model.VssSpecification
+import org.eclipse.kuksa.vsscore.model.model.className
+import org.eclipse.kuksa.vsscore.model.model.name
+import org.eclipse.kuksa.vsscore.model.model.variableName
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
 import kotlin.reflect.full.declaredMemberProperties
